refactor(core-weave): extract payload render helpers
- move first and second payload KnopInventory Turtle renderers into payload_renderers.ts
- keep multi-history payload rendering private to the payload renderer module
- move shared support-history omission postprocessors into support_history_renderers.ts
- keep mesh-inventory renderers, planner dispatch, and generated RDF/Page output unchanged
